Add library data via CI/CD
This PR adds the library to the site for use within use cases.  The
library is created within https://github.com/inngest/library;  on merges
to main, we create a new release with a JSON file containing all library
data.

We fetch this data in two ways:

1. Statically, at prod build time, via a makefile command (`make
   library`).  This downloads the latest library JSON file into
   `./public/json/library.json`, which can be fetched as a public API.
2. Dynamically, _in dev_.  We create a new `utils/library.js` fetcher
   which downloads the latest library JSON file automatically in
   development, ensuring that we have the latest data when developing
   the website.

**Usage**

We create a new `./utils/library.ts` file which exports a single async
function.  This async function:

1. Fetches the library (using pre-downloaded data in prod, or using
   latest data in dev).
2. Initializes a new `LibraryManager` class with the library data

We use this as follows:

```
import getLibrary from "utils/library"

export async function getStaticProps(context) {
  // This fetches the library via async functionality on the server before rendering,
  // letting us dynamically generate the use cases.
  const library = await getLibrary();

  // We can now use library as a `LibraryManager`:

  return {
    props: {
      useCases: library.forCategory("webhooks"),
    },
  }
}
```

This passes use cases for a specific category into the page at build
time.