Update curl command in cloud.mdx (#1337)
Add --fail-with-body option to curl command for PUT request.

Quality of life suggestion for the documentation: I myself ran into this issue when adding the CURL command to my Github Actions and the URL was subtly wrong. The deployment did continue without an error but Inngest didn't update because the CURL command failed silently.